pyproject.toml
[project]
name = "telegram-supervisor"
version = "0.1.0"
description = "Telegram bot for moderating educational chats."
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "Anda4ka"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aiogram>=3.5",
    "alembic>=1.13",
    "psycopg2-binary>=2.9",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "python-dotenv",
    "structlog>=23.2",
    "pytz>=2024.1",
    "pydantic-ai[openai]>=1.66,<2.0",
    "httpx>=0.27",
    "feedparser>=6.0.12",
    "telethon>=1.42.0",
    "burr>=0.40.2",
    "telegramify-markdown==1.0.0rc5",
    "pgvector>=0.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
src = ["app", "tests"]
target-version = "py312"
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",   # pep8-naming
    "S",   # flake8-bandit
    "T20", # flake8-print
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
]
ignore = [
    "E501",  # line too long
    "S101",  # use of assert
    "T201",  # print found
    "B008",  # function call in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "ARG001",  # unused function arguments (common in tests/fixtures)
    "ARG002",  # unused method arguments (common in test methods)
    "S311",    # pseudo-random generators (acceptable in tests)
    "SIM117",  # nested with statements (common in test assertions)
    "E402",    # module level import not at top (pytest fixtures)
    "N815",    # mixed case variable name (test data)
    "N802",    # camelCase method names (matching Telegram API methods in fakes)
    "N806",    # uppercase variable names in functions (test constants)
    "S106",    # hardcoded passwords (test tokens)
    "S110",    # try-except-pass (docker availability check)
    "PLR2004", # magic value used in comparison
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# ── ty type checker (replaces mypy) ──────────────────────────────────
[tool.ty.environment]
python-version = "3.12"

[tool.ty.src]
include = ["app", "tests"]
exclude = ["alembic/**"]

[tool.ty.rules]
# Third-party stubs often missing or incomplete
unresolved-import = "warn"
# Optional narrowing false positives (e.g., telethon_wrapper.is_available after None check)
unresolved-attribute = "warn"

[tool.ty.analysis]
# Telethon, burr, feedparser, pydantic-ai lack complete stubs
replace-imports-with-any = [
    "telethon.**",
    "burr.**",
    "feedparser.**",
    "testcontainers.**",
]

# Agent + moderation modules — PydanticAI dynamic patterns, framework constructors
[[tool.ty.overrides]]
include = ["app/agent/**", "app/moderation/**"]

[tool.ty.overrides.rules]
invalid-argument-type = "warn"       # PydanticAI/Pydantic constructor patterns
invalid-return-type = "warn"         # dynamic returns from agent frameworks
invalid-assignment = "warn"          # PydanticAI agent type variance

# Core — DI container generic patterns, Pydantic config
[[tool.ty.overrides]]
include = ["app/core/**"]

[tool.ty.overrides.rules]
invalid-return-type = "warn"         # generic container return patterns
invalid-argument-type = "warn"       # Pydantic Field patterns

# Tests — relaxed checking for mocks, fixtures, test data
[[tool.ty.overrides]]
include = ["tests/**"]

[tool.ty.overrides.rules]
invalid-argument-type = "warn"       # mock/fixture type mismatches
invalid-return-type = "warn"         # test helpers
no-matching-overload = "warn"        # mock constructor patterns

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
    "aiosqlite>=0.20",
    "ruff>=0.12.4",
    "ty>=0.0.23",
    "pre-commit>=3.6",
    "watchfiles>=0.21",
    "psutil",
    "testcontainers[postgres]>=4.14.1",
]

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "e2e: marks tests as end-to-end tests",
    "performance: marks tests as performance tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__main__.py",
    "*/migrations/*",
]

[tool.coverage.report]
show_missing = true
fail_under = 45
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
